Add ability to deploy contracts non-interactively
